improve equipment logic

+++ src/main/java/com/jftse/emulator/server/core/handler/game/inventory/InventoryItemTimeExpiredPacketHandler.java
@@ -7,6 +7,7 @@ import com.jftse.emulator.server.core.packet.packets.inventory.C2SInventoryItemT
 import com.jftse.emulator.server.core.packet.packets.inventory.S2CInventoryItemRemoveAnswerPacket;
 import com.jftse.emulator.server.core.service.PlayerPocketService;
 import com.jftse.emulator.server.core.service.PocketService;
+import com.jftse.emulator.server.database.model.player.Player;
 import com.jftse.emulator.server.database.model.pocket.PlayerPocket;
 import com.jftse.emulator.server.database.model.pocket.Pocket;
 import com.jftse.emulator.server.networking.packet.Packet;
@@ -32,17 +33,22 @@ public class InventoryItemTimeExpiredPacketHandler extends AbstractHandler {
 
     @Override
     public void handle() {
+        if (connection.getClient() == null || connection.getClient().getPlayer() == null)
+            return;
+
+        Player player = connection.getClient().getPlayer();
+        Pocket pocket = player.getPocket();
+
         long itemPocketId = inventoryItemTimeExpiredReqPacket.getItemPocketId();
-        PlayerPocket item = playerPocketService.findById(itemPocketId);
+        PlayerPocket item = playerPocketService.getItemAsPocket(itemPocketId, pocket);
         if (item != null && item.getUseType().equals(EItemUseType.TIME.getName())) {
             long timeLeft = (item.getCreated().getTime() * 10000) - (new Date().getTime() * 10000);
             if (timeLeft < 0) {
                 playerPocketService.remove(itemPocketId);
 
-                Pocket pocket = connection.getClient().getPlayer().getPocket();
                 pocketService.decrementPocketBelongings(pocket);
 
-                S2CInventoryItemRemoveAnswerPacket inventoryItemRemoveAnswerPacket = new S2CInventoryItemRemoveAnswerPacket(inventoryItemTimeExpiredReqPacket.getItemPocketId());
+                S2CInventoryItemRemoveAnswerPacket inventoryItemRemoveAnswerPacket = new S2CInventoryItemRemoveAnswerPacket(item.getId().intValue());
                 connection.sendTCP(inventoryItemRemoveAnswerPacket);
             }
         }
